DCMAW-8264: Move types around for co-location
